+++ src/main/java/com/kakaotrack/choco/linkupapi/user/UserRepository.java
@@ -4,9 +4,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<SiteUser, Long> {
+public interface UserRepository extends JpaRepository<SiteUser, Integer> {
     Optional<SiteUser> findByusername(String username);
 
-    Optional<SiteUser> findById(Long id);
+    Optional<SiteUser> findById(int id);
 
 }
